requirements.txt
py-vox-io>=0.1,<1.0
Pillow>=10.0,<12.0